TheHiveApiExtended is considered as an action and crash on https://github.com/StackStorm/st2/blob/master/st2common/st2common/util/loader.py#L69
