+++ actions/run_analyzer_on_data_type.py
@@ -1,17 +1,7 @@
 import random
-import requests
-from thehive4py.api import TheHiveApi
+from thehive4pyextended import TheHiveApiExtended
 from thehive4py.query import Eq
 from st2common.runners.base_action import Action
-
-
-class TheHiveApiExtended(TheHiveApi):
-    def get_analyzer_by_name_and_data_type(self, name, data_type):
-        req = self.url + "/api/connector/cortex/analyzer/type/{}".format(data_type)
-        analyzers = requests.get(req, proxies=self.proxies, auth=self.auth, verify=self.cert)
-        analyzer = [a for a in analyzers.json() if a['name'].lower() == name.lower()]
-        return analyzer[0]
-
 
 __all__ = [
     'RunAnalyzerOnDataTypeAction'
